pyproject.toml
[project]
name = "healthpay-assignment-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.13",
    "loguru>=0.7.3",
    "mistralai>=1.8.2",
    "pydantic-settings>=2.9.1",
    "uvicorn[standard]>=0.34.3",
]

[dependency-groups]
dev = [
    "ruff>=0.12.0",
]


[tool.ruff]
line-length = 150
target-version = "py313"

[tool.ruff.isort]
known-first-party = ["app"]
known-third-party = ["fastapi", "pydantic", "langchain"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
]
